react update
- 상세페이지 지도 병합
